+++ src/components/DeploymentsTable.vue
@@ -53,7 +53,7 @@
       <template #action="{ row }">
         <div class="deployments-table__actions">
           <DeploymentToggle :deployment="row" @update="refresh" />
-          <DeploymentMenu size="xs" :deployment="row" @delete="refresh" />
+          <DeploymentMenu size="xs" :deployment="row" show-all @delete="refresh" />
         </div>
       </template>
 
@@ -78,7 +78,6 @@
   import { useSubscription } from '@prefecthq/vue-compositions'
   import { computed, unref, ref } from 'vue'
   import { SearchInput, ResultsCount, DeploymentToggle, DeploymentMenu, FlowRouterLink, FlowCombobox, DeploymentsDeleteButton, SelectedCount } from '@/components'
-
   import { useWorkspaceApi, useWorkspaceRoutes, useCan } from '@/compositions'
   import { UseDeploymentFilterArgs, useDeploymentFilterFromRoute } from '@/compositions/useDeploymentFilter'
   import { isRRuleSchedule, Schedule } from '@/models'
@@ -215,5 +214,6 @@
 .deployments-table__actions { @apply
   flex
   gap-2
+  items-center
 }
 </style>
