+++ src/components/index.ts
@@ -38,6 +38,7 @@ export { default as FlowRunsDeleteButton } from './FlowRunsDeleteButton.vue'
 export { default as FlowsDeleteButton } from './FlowsDeleteButton.vue'
 export { default as DeploymentCombobox } from './DeploymentCombobox.vue'
 export { default as DeploymentComboboxOption } from './DeploymentComboboxOption.vue'
+export { default as DeploymentCustomRunOverflowMenuItem } from './DeploymentCustomRunOverflowMenuItem.vue'
 export { default as DeploymentDeprecatedMessage } from './DeploymentDeprecatedMessage.vue'
 export { default as DeploymentDescription } from './DeploymentDescription.vue'
 export { default as DeploymentDescriptionEmptyState } from './DeploymentDescriptionEmptyState.vue'
@@ -45,6 +46,7 @@ export { default as DeploymentDetails } from './DeploymentDetails.vue'
 export { default as DeploymentForm } from './DeploymentForm.vue'
 export { default as DeploymentIconText } from './DeploymentIconText.vue'
 export { default as DeploymentMenu } from './DeploymentMenu.vue'
+export { default as DeploymentQuickRunOverflowMenuItem } from './DeploymentQuickRunOverflowMenuItem.vue'
 export { default as DeploymentsCount } from './DeploymentsCount.vue'
 export { default as DeploymentsDeleteButton } from './DeploymentsDeleteButton.vue'
 export { default as DeploymentsPageEmptyState } from './DeploymentsPageEmptyState.vue'
